Fix tracing when running in AWS Lambda
The AWS Lambda infrastructure automatically creates a top-level trace
Segment for the duration of the lambda's execution; however client code
does not have access to this, and it does *not* appear on the main
recorder object in the usual way. This messes with our automatic
detection of parent-child relationships.

Instead, try to detect this case and forcibly call beginSubsegment()
when running in lambda.