Allow in-progress segments to be sent to X-Ray on start()
The default behaviour is to only send trace data back to X-Ray once the
entire trace tree is finished. However, it can be useful to send some
in-progress segments back to X-Ray as soon as they start so we can
monitor their progress.

This should be used with caution, since it involves streaming data back
to the X-Ray servers (depending on the Emitter configured).