Add AWSXRayTags.PARENT_ID to directly call Entity.setParentId()
By default, the parentId is _not_ set when a child Entity is generated:
this is redundant since, if the whole trace tree is serialized back to
X-Ray at once, the parent-child relationships are inherent in the JSON
structure.

However in the case where parts of the trace are being streamed
separately, or sent back early while still in-progress, we need to
set the parentId to ensure the final tree can be reconstructed properly.